snapshot commit before changing things up.
in this snapshot the wavesetup data is an input and
this version will closest neighbour apply the wave setup
value at a location to the closest location on the roms coast.

I'm going to change this so that the setup is calculated directly
on the roms coast. I will now read in the AUSWAVE Hs and Tp data,
natural neighbour interpolate the field, then directly calculate the
setup component on the roms coastal points.